Feat: 메서드를 분리하는 건 단일 책임을 위해선 좋지만 연관이 되어있는 메서드들이라면 예외를 적용하라

+++ src/main/java/nextstep/qna/domain/Answer.java
@@ -60,7 +60,10 @@ public class Answer {
         return this.writer.equals(writer);
     }
 
-    public void addTo(List<DeleteHistory> deleteHistories) {
+    public void addTo(List<DeleteHistory> deleteHistories) throws CannotDeleteException {
+        if (!deleted) {
+            throw new CannotDeleteException("답변이 삭제가 되지 않았습니다");
+        }
         deleteHistories.add(new DeleteHistory(ContentType.ANSWER, id, writer, LocalDateTime.now()));
     }
